+++ next.config.ts
@@ -2,6 +2,12 @@ import type { NextConfig } from "next";
 
 const nextConfig: NextConfig = {
   /* config options here */
+  images: {
+    domains: [
+      "lh3.googleusercontent.com",
+      // add any other domains you need here
+    ],
+  },
 };
 
 export default nextConfig;